Added concurrently to start both servers at the same time.
