+++ Peer-Connect-Backend/routes/messageRoute.js
@@ -1,10 +1,8 @@
-const express = require('express');
-const { User, Message, Conversation } = require('../db');
-const { sendMessage } = require('../controllers/messageController');
-const checkSession = require('../middlewares/sessionAuth');
+const express = require("express");
+const { sendMessage } = require("../controllers/messageController");
+const checkSession = require("../middlewares/sessionAuth");
 const router = express.Router();
 
-
-router.post("/send", checkSession ,sendMessage);
+router.post("/send", checkSession, sendMessage);
 
 module.exports = router;
